requirements.txt
aiofiles
aiohttp
aiosignal
annotated-types
anyio
asgiref
async-timeout
attrs
certifi
charset-normalizer
click
dataclasses-json
Deprecated
dirtyjson
distro
Django
exceptiongroup
frozenlist
fsspec
greenlet
h11
httpcore
httpx
idna
joblib
jsonpatch
jsonpointer
langchain
langchain-core
langchain-openai
langchain-text-splitters
langchain-together
langsmith
llama-index-core
llama-parse
llamaindex-py-client
marshmallow
multidict
mypy-extensions
nest-asyncio
networkx
nltk
numpy
openai
orjson
packaging
pandas
pillow
pydantic
pydantic_core
python-dateutil
python-dotenv
pytz
PyYAML
regex
requests
six
sniffio
SQLAlchemy
sqlparse
tenacity
umap-learn
tiktoken
scipy
torch
transformers
scikit-learn
tqdm
python-magic
RAGatouille
Markdown
beautifulsoup4
pypandoc
FlagEmbedding
unstructured
gptpdf
pymupdf4llm
azure-ai-documentintelligence